// API/schemas.ts
export const cartUpsertSchema = z.object({
  storeId: z.string().uuid(),
  variantId: z.string().uuid(),
  quantity: z.coerce.number().int().positive(),
});

export const cartRemoveSchema = z.object({
  storeId: z.string().uuid(),
  variantId: z.string().uuid(),
});